pyproject.toml
[tool.poetry]
name = "gpx-track-render-er"
version = "0.1.0"
description = ""
authors = ["Ieuan Jenkins <hi@ieuan.net>"]

[tool.poetry.dependencies]
python = "^3.9"
gpxpy = "^1.4.2"
geopy = "^2.1.0"
Faker = "^5.1.0"
pytz = "^2020.5"
PyYAML = "^5.3.1"
confuse = "^1.4.0"
isodate = "^0.6.0"
"SRTM.py" = "^0.3.6"
mailbox = "^0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
